Refactor farming logic out of `SingleDiskPlot` into `farming` module
